requirements.txt
aiofiles==0.5.0
apache-beam==2.20.0
attrs==19.3.0
# avro-python3==1.9.2.1
awesome-slugify==1.6.5
bitarray==1.2.2
bitstring==3.1.7
blessings==1.7
blinker==1.4
boto3==1.13.12
botocore==1.16.12
cached-property==1.5.1
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
crcmod==1.7
dataflows==0.0.72
datapackage==1.13.1
dill==0.3.1.1
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
fastavro==0.21.24
fasteners==0.15
Flask==1.1.2
flask-wtf
flask-googlemaps==0.4.0
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
futures==3.1.1
google-api-core==1.16.0
google-api-python-client==1.8.3
google-apitools==0.5.28
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-automl==0.10.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-pubsub==0.20.0
google-cloud-storage
google-cloud-translate==2.0.1
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.10.0
gunicorn==19.7.1
grpcio==1.28.1
grpcio-gcp==0.2.2
h11==0.9.0
h2==3.2.0
hdfs==2.5.8
hpack==3.0.0
httplib2==0.12.0
Hypercorn==0.9.5
hyperframe==5.2.0
idna==2.9
ijson==3.0.4
importlib-metadata==1.6.0
inquirer==2.6.3
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jsonlines==1.2.0
jsonpointer==2.0
jsonschema==3.2.0
kvfile==0.0.7
linear-tsv==1.1.0
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.4
openpyxl==3.0.3
pbr==5.4.5
priority==1.3.0
protobuf==3.11.3
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybloom-live==3.0.0
pydot==1.4.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
Quart==0.11.5
readchar==2.0.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.17
tableschema==1.17.2
tableschema-sql==1.3.1
tabulate==0.8.7
tabulator==1.46.0
toml==0.10.1
tweepy==3.8.0
typing-extensions==3.7.4.2
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wsproto==0.15.0
xlrd==1.2.0
xmljson==0.2.1
zipp==3.1.0